backend/requirements.txt
asgiref==3.9.1
django==5.2.6
django-cors-headers==4.7.0
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
django-storages
boto3
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pdfminer.six==20250327
pdfplumber==0.11.6
protobuf==3.20.3
pypdfium2==4.30.1
python-dotenv==1.1.1
setuptools==78.1.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
wheel==0.45.1
